Ngx Translate universal fix
