+++ src/app/browser-app.module.ts
@@ -4,6 +4,13 @@ import { AppComponent } from './app.component';
 import { AppModule } from './app.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ServiceWorkerModule } from '@angular/service-worker';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+
+export function HttpLoaderFactory(http: HttpClient) {
+    return new TranslateHttpLoader(http);
+}
 
 @NgModule({
     bootstrap: [ AppComponent ],
@@ -13,7 +20,11 @@ import { ServiceWorkerModule } from '@angular/service-worker';
             appId: 'app'
         }),
         AppModule,
-        ServiceWorkerModule.register('/ngsw-worker.js')
+        ServiceWorkerModule.register('/ngsw-worker.js'),
+        HttpClientModule,
+        TranslateModule.forRoot({
+            loader: {provide: TranslateLoader, useFactory: HttpLoaderFactory, deps: [HttpClient]}
+        })
     ]
 })
 export class BrowserAppModule {}
